# backend-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Share Backend API",
    "description": "Share Backend API",
    "version": "v0.1.0"
  },
  "servers": [
    {
      "url": "https://hub.maayuan.fun:16666",
      "description": "Public HTTPS endpoint"
    }
  ],
...
  "paths": {
...
    "/v1/inventory/import": {
      "post": {
        "tags": [
          "库存"
        ],
        "summary": "导入库存/奖励",
        "operationId": "import",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InventoryImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultInventoryImportResult"
                }
              }
            }
          },
          "400": {
            "description": "invalid_json",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "account_not_found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "record_conflict or account_name_conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "schema_validation_failed, unknown_account_id, unknown_entity_id, or unsupported_version",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Jwt": []
          }
        ]
      }
    },
    "/v1/inventory/accounts": {
      "get": {
        "tags": [
          "库存"
        ],
        "summary": "库存子账号列表",
        "operationId": "accounts",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultListInventoryAccountResponse"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "account_not_found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "schema_validation_failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Jwt": []
          }
        ]
      },
      "post": {
        "tags": [
          "库存"
        ],
        "summary": "创建库存子账号",
        "operationId": "createAccount",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InventoryAccountRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultInventoryAccountResponse"
                }
              }
            }
          },
          "400": {
            "description": "invalid_json",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "account_not_found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "record_conflict or account_name_conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "schema_validation_failed, unknown_account_id, unknown_entity_id, or unsupported_version",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Jwt": []
          }
        ]
      }
    },
...
    "/open-api/inventory/import": {
      "post": {
        "tags": [
          "OpenAPI Inventory"
        ],
        "summary": "导入库存/奖励",
        "operationId": "inventoryImport",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InventoryImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultInventoryImportResult"
                }
              }
            }
          },
          "400": {
            "description": "invalid_json",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "account_not_found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "record_conflict or account_name_conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "schema_validation_failed, unknown_account_id, unknown_entity_id, or unsupported_version",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OpenApiToken": []
          }
        ]
      }
    },
...
    "/v1/inventory/accounts/{accountId}": {
      "delete": {
        "tags": [
          "库存"
        ],
        "summary": "删除库存子账号及其库存和流水",
        "operationId": "deleteAccount",
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultBoolean"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "record_not_found or account_not_found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Jwt": []
          }
        ]
      },
      "patch": {
        "tags": [
          "库存"
        ],
        "summary": "修改库存子账号名称",
        "operationId": "renameAccount",
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InventoryAccountRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultInventoryAccountResponse"
                }
              }
            }
          },
          "400": {
            "description": "invalid_json",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "account_not_found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "record_conflict or account_name_conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "schema_validation_failed, unknown_account_id, unknown_entity_id, or unsupported_version",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Jwt": []
          }
        ]
      }
    },
...
    "/v1/inventory/records": {
      "get": {
        "tags": [
          "库存"
        ],
        "summary": "导入记录列表",
        "description": "导入记录列表(需登录;entity_type/from/to 可选过滤,按 effective_at 倒序)。",
        "operationId": "records",
        "parameters": [
          {
            "name": "account_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "entity_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "item",
                "agent"
              ]
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50,
              "maximum": 100,
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultInventoryRecordPageResponse"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "account_not_found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "schema_validation_failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Jwt": []
          }
        ]
      }
    },
    "/v1/inventory/export": {
      "get": {
        "tags": [
          "库存"
        ],
        "summary": "导出",
        "description": "导出(需登录;默认仅当前状态 full 快照,include=rewards 时附带区间奖励流水)。",
        "operationId": "export",
        "parameters": [
          {
            "name": "account_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "all"
              ]
            }
          },
          {
            "name": "include",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "current",
                "current,rewards"
              ]
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryExportResponse"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "account_not_found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "schema_validation_failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "Jwt": []
          }
        ]
      }
    },
...
    "/v1/inventory/catalog": {
      "get": {
        "tags": [
          "库存"
        ],
        "summary": "对象目录",
        "description": "对象目录(公开,无需登录;返回目录版本与全部 {entity_type, id, name})。",
        "operationId": "catalog",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultInventoryCatalogResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "InventoryImportRequest": {
        "properties": {
          "format": {
            "type": "string",
            "enum": [
              "myshare-inventory-exchange"
            ],
            "minLength": 1,
            "pattern": "^myshare-inventory-exchange$"
          },
          "version": {
            "type": "integer",
            "format": "int32",
            "maximum": 2,
            "minimum": 2
          },
          "exported_at": {
            "type": "string",
            "format": "date-time",
            "minLength": 1
          },
          "producer": {
            "$ref": "#/components/schemas/ProducerDto"
          },
          "catalog_version": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1
          },
          "accounts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InventoryExchangeAccountDto"
            },
            "maxItems": 1000,
            "minItems": 1
          },
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InventoryRecordRequest"
            },
            "maxItems": 1000,
            "minItems": 1
          }
        },
        "required": [
          "exported_at",
          "format",
          "producer",
          "records",
          "version"
        ]
      },
...
      "InventoryErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/InventoryError"
          }
        },
        "required": [
          "error"
        ]
      },
      "ApiResultInventoryImportResult": {
        "properties": {
          "status_code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/InventoryImportResult"
          }
        },
        "required": [
          "status_code"
        ]
      },
...
      "ApiResultListInventoryAccountResponse": {
        "properties": {
          "status_code": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InventoryAccountResponse"
            }
          }
        },
        "required": [
          "status_code"
        ]
      },
...
    },
    "securitySchemes": {
      "Jwt": {
        "type": "http",
        "name": "Authorization",
        "in": "header",
        "scheme": "bearer"
      },
      "OpenApiToken": {
        "type": "http",
        "name": "Authorization",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "API token"
      }
    }
  }
}